Fix ci failures
